refactor: break up DB classes into focused repositories
